requirements.txt
# Core
pandas==2.2.0
numpy==1.26.4
scipy==1.12.0

# Geospatial
geopandas==0.14.3
folium==0.16.0
shapely==2.0.3
pyproj==3.6.1
osmnx==1.9.1

# Visualization
plotly==5.20.0
matplotlib==3.8.3
seaborn==0.13.2

# Web Scraping
requests==2.31.0
beautifulsoup4==4.12.3
lxml==5.1.0
httpx==0.27.0

# Bayesian Modeling
pymc==5.10.4
arviz==0.18.0

# Notebooks
jupyter==1.0.0
jupyterlab==4.1.2
ipywidgets==8.1.2

# Utilities
python-dotenv==1.0.1
tqdm==4.66.2
pyyaml==6.0.1

# Testing
pytest==8.0.2
pytest-cov==4.1.0
